* merge of new_output branch rev 1810:1876

+++ include/core/macroheatbalance/MacroHeatBalance.h
@@ -23,7 +23,7 @@ class MeshBase;
 //!  Class to solve heat transport problem
 class MacroHeatBalance : public SimulationInterface
 /*!
-  
+
  \f$ -\nabla_i \cdot(k_{i,j} \nabla_j T)+ \nabla \cdot (T P_n J_n + T P_p J_p)=-\nabla cdot (J_n \phi_n + J_p \phi_p)\f$
  * The get_solution() methods can provide the following variables:
  * Temperature (K)
@@ -42,7 +42,7 @@ class MacroHeatBalance : public SimulationInterface
      JQY,
      JQZ
    };
-       
+
 
 
 
@@ -50,10 +50,10 @@ class MacroHeatBalance : public SimulationInterface
   //!options that we need for this simulation
   struct options
   {
-  
+
     std::string  kappa_solve; //!< Model for lattice thermal conductivity
-      
-    double max_error; //!< Max tollerance for self-consistent loop  
+
+    double max_error; //!< Max tollerance for self-consistent loop
 
     double work_units; //!< SI units, has to be consistent with the database parameters
 
@@ -67,7 +67,7 @@ class MacroHeatBalance : public SimulationInterface
 
   };
 
-    
+
   /*!
    * \copydoc SimulationInterface::get_solution_secure(const Elem*,
    * const std::vector<Point>&, const std::vector<ID>&,
@@ -75,39 +75,39 @@ class MacroHeatBalance : public SimulationInterface
    */
   virtual void get_solution_secure(const Elem* elem, const std::vector<Point>& p,
 				   const std::set<ID>& ids, std::vector<std::map<ID, double> >& values);
-  
-  
+
+
   /*!
    * \copydoc SimulationInterface::get_solution_secure(const Elem*,
    * const std::vector<ID>&, std::vector<std::vector<double> >&)
-   */  
+   */
   virtual void get_solution_secure(const Elem* elem,
 				   const std::set<ID>& ids, std::vector<std::map<ID, double> >& values);
-  
+
 
   //!Constructor
   MacroHeatBalance(const ModelOptions& options);
-  
+
   //!Destructor
   virtual ~MacroHeatBalance();
-  
+
   virtual PhysicalModel* create_physical_model(const ModelOptions &options,
       const Material* mat) const throw (ModelErrorException);
-  
-  
-  virtual BoundaryProperties* create_boundary_model(const ModelOptions &options) const 
+
+
+  virtual BoundaryProperties* create_boundary_model(const ModelOptions &options) const
     throw (ModelErrorException);
 
   static void assemble_heat_matrix(EquationSystems& es,
 				     const std::string& system_name);
 
-  //!Create an MacroHeatBalance object 
+  //!Create an MacroHeatBalance object
   static MacroHeatBalance*  create(const ModelOptions& options);
-  
+
   /*! \copydoc SimulationInterface::build_integrated_quantities() */
   virtual void build_integrated_quantities(std::vector<double>& values);
-  
-  
+
+
   /*! \copydoc SimulationInterface::build_integrated_quantities_description()
    */
   virtual void build_integrated_quantities_description(
@@ -118,66 +118,66 @@ class MacroHeatBalance : public SimulationInterface
     virtual void do_print_info(void);
 
  private:
-  
+
   std::string heat_legend;
 
   std::set<ID> JQ_var;
 
-   //! Quadrature point along the face of the element 
+   //! Quadrature point along the face of the element
   const std::vector<Point> qface_point;
 
   //! A pointer to heat simulation
   MacroHeatBalance* _heat_simul;
 
   EquationSystems * 	equation_systems;
-  
-  HeatModel* heat_model; 
+
+  HeatModel* heat_model;
 
   //void init_heat_model(const Elem* elem);
- 
+
 
   std::string system_name;
-  
-  TiberLinearSystem* my_system;  
- 
+
+  TiberLinearSystem* my_system;
+
   //! Order the solution in correct mode
-  virtual void 	build_nodal_results(const std::set< std::string > &variables, 
-				     std::vector< double > &results, 
+  virtual void 	build_nodal_results(const std::set< std::string > &variables,
+				     std::vector< double > &results,
 				     std::vector< std::string > &legend);
 
-   
+
   //! Order the solution in correct mode
   virtual void build_elemental_results(const std::set<std::string>& variables,
-				       std::vector<double>& results, 
+				       std::vector<double>& results,
 				       std::vector<std::string>& legend);
 
-  //! Calculate Power Dissipated 
+  //! Calculate Power Dissipated
   /*!
    * Integrates numerically over the boundary elements.
    * The power dissipated is then:
    *
    * \f[P = \int_{\Gamma} -\kappa \nabla T \cdot \mathbf{N} \mathrm{d}\Gamma \f]
    */
-    
+
   double calculate_power_emitted(void);
 
   double calculate_power_dissipated_rstf(void);
 
-  double calculate_power_dissipated(void); 
-  
+  double calculate_power_dissipated(void);
+
   static MacroHeatBalance* static_this;
 
   options myopts;
 
-  //!non-static method that actually does matrix assembling 
+  //!non-static method that actually does matrix assembling
   void do_assemble(EquationSystems& es, const std::string& system_name);
 
   static Device* _device;
 
   //!Dimension of meshmap
-  short dim;  
+  unsigned int dim;
 
-  
+
 
   //!Pointer to mesh
   MeshBase* mesh;
@@ -186,17 +186,17 @@ class MacroHeatBalance : public SimulationInterface
 
     /*! \copydoc SimulationInterface::convert_variable_name_to_id() */
   virtual ID convert_variable_name_to_id(const std::string& variable_name) const;
- 
-  //! \copydoc  SimulationInterface::do_init() 
+
+  //! \copydoc  SimulationInterface::do_init()
   virtual void 	do_init (void);
-  
+
   //!Do the solve
   virtual void do_solve (void);
- 
+
   //!Parse the options
   virtual void 	parse_options(void);
- 
+
 
 };
- 
+
 #endif
